feat: widen interpreter array length/index to ECMA-262 uint32 (#73 Stage E.1)
SharpTSArray._length is now long (was int). All Set/Get/HasIndex/GetRaw/
DeleteAt/SetLength paths have long overloads; int callers implicitly widen.
IndexTarget.Array.Index → long so ResolveIndexTarget handles double indices
up to 2^32-1 without wrapping. JS `length` property accessor reads through
LongLength, so arr.length === 4294967295 is now representable.

Public int Length clamps at int.MaxValue for source compatibility (10k+
call sites do `for (int i = 0; i < arr.Length; i++)` — widening the
public property would break them). LongLength exposes the true value.

Test262 wins (interpreter, RuntimeError → Pass):
 - test/built-ins/Array/S15.4.5.2_A1_T1.js  (issue #73 canary)
 - test/built-ins/Array/S15.4.5.2_A3_T3.js
 - test/built-ins/Array/length/15.4.5.1-3.d-3.js
 - test/built-ins/Array/length/S15.4.5.2_A3_T4.js

10,275 SharpTS.Tests still green.

Compile-mode parity (E.2) not included — compile-mode arrays still
use List<object?> with int-bounded indices and no hole sentinel. That
port is a separate undertaking.

+++ Execution/IndexTarget.cs
@@ -11,7 +11,12 @@ public abstract record IndexTarget
     private IndexTarget() { }
 
     // Get/Set targets
-    public sealed record Array(SharpTSArray Target, int Index) : IndexTarget;
+    /// <summary>
+    /// Array index target — <c>Index</c> is <c>long</c> so JS literals like
+    /// <c>a[2147483648]</c> (which exceed int range but sit inside ECMA-262's
+    /// uint32 array index range) route correctly through SharpTSArray's long APIs.
+    /// </summary>
+    public sealed record Array(SharpTSArray Target, long Index) : IndexTarget;
     public sealed record TypedArray(SharpTSTypedArray Target, int Index) : IndexTarget;
     public sealed record Buffer(SharpTSBuffer Target, int Index) : IndexTarget;
     public sealed record ObjectString(SharpTSObject Target, string Key) : IndexTarget;
